perf:新建useGetChatDataStoreKey
